ai-services/analysis-service/app.py
from fastapi import FastAPI, File, UploadFile, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional
import numpy as np
import cv2
from io import BytesIO
from PIL import Image

app = FastAPI(title="FotoAI Analysis Service", version="1.0.0")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class BoundingBox(BaseModel):
    x: int
    y: int
    width: int
    height: int

class PrivacyObject(BaseModel):
    type: str
    bbox: BoundingBox
    confidence: float
    suggested_blur: bool

class RecommendedAdjustments(BaseModel):
    light_mood: float
    air: float
    depth: float
    presence: float
    detail: float

class AnalysisResponse(BaseModel):
    scene_type: str
    lighting_quality: float
    color_temp_kelvin: int
    dynamic_range_stops: float
    has_sky: bool
    reflective_surfaces: List[str]
    privacy_objects: List[PrivacyObject]
    recommended_adjustments: RecommendedAdjustments

@app.get("/health")
async def health_check():
    return {"status": "healthy", "service": "analysis"}

@app.post("/analyze", response_model=AnalysisResponse)
async def analyze_image(file: UploadFile = File(...)):
    """
    Analyze uploaded image and return comprehensive scene analysis
    """
    try:
        # Read image
        contents = await file.read()
        image = Image.open(BytesIO(contents))
        img_array = np.array(image)
        
        # TODO: Implement actual AI models
        # - Scene classification (interior/exterior, room type)
        # - Lighting analysis
        # - Privacy object detection (windows, reflections)
        # - Dynamic range estimation
        
        # Placeholder response
        analysis = AnalysisResponse(
            scene_type="interior_living_room",
            lighting_quality=0.72,
            color_temp_kelvin=4200,
            dynamic_range_stops=8.5,
            has_sky=False,
            reflective_surfaces=["window", "glass_table"],
            privacy_objects=[
                PrivacyObject(
                    type="window_view",
                    bbox=BoundingBox(x=120, y=80, width=220, height=200),
                    confidence=0.91,
                    suggested_blur=True
                )
            ],
            recommended_adjustments=RecommendedAdjustments(
                light_mood=0.15,
                air=0.22,
                depth=-0.08,
                presence=0.31,
                detail=0.12
            )
        )
        
        return analysis
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Analysis failed: {str(e)}")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8001)
