ai-services/seo-service/app.py
from fastapi import FastAPI, File, UploadFile, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Optional, Dict, Any
from PIL import Image
from io import BytesIO

app = FastAPI(title="FotoAI SEO Service", version="1.0.0")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class SEOMetadata(BaseModel):
    filename: str
    alt_text: str
    caption: str
    structured_data: Dict[str, Any]

class SEORequest(BaseModel):
    address: Optional[str] = None
    listing_type: Optional[str] = "residential"
    room_type: Optional[str] = None

@app.get("/health")
async def health_check():
    return {"status": "healthy", "service": "seo"}

@app.post("/generate-seo", response_model=SEOMetadata)
async def generate_seo(
    file: UploadFile = File(...),
    metadata: Optional[str] = None
):
    """
    Generate SEO-optimized metadata for real estate photos
    """
    try:
        # Parse metadata
        if metadata:
            import json
            meta = SEORequest(**json.loads(metadata))
        else:
            meta = SEORequest()
        
        # Read image
        contents = await file.read()
        image = Image.open(BytesIO(contents))
        
        # TODO: Implement vision-language model (CLIP/BLIP-2)
        # - Analyze image content
        # - Generate descriptive text
        # - Create SEO-friendly filename
        
        # Placeholder generation
        room_type = meta.room_type or "living-room"
        address_slug = meta.address.replace(" ", "-").lower() if meta.address else "property"
        
        filename = f"modern-{room_type}-natural-light-{address_slug}.jpg"
        
        alt_text = f"Spacious modern {room_type.replace('-', ' ')} with natural light"
        if meta.address:
            alt_text += f" at {meta.address}"
        
        caption = f"Bright, open-concept {room_type.replace('-', ' ')} featuring abundant natural light"
        
        structured_data = {
            "@context": "https://schema.org",
            "@type": "ImageObject",
            "contentUrl": "https://cdn.fotoai.com/placeholder",
            "description": alt_text,
            "about": {
                "@type": "RealEstateListing",
                "address": meta.address or "Property Address"
            }
        }
        
        return SEOMetadata(
            filename=filename,
            alt_text=alt_text,
            caption=caption,
            structured_data=structured_data
        )
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"SEO generation failed: {str(e)}")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8004)
