ai-services/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Image Processing
opencv-python==4.9.0.80
pillow==10.2.0
imageio==2.33.1
rawpy==0.19.1
numpy==1.26.3
scikit-image==0.22.0

# AI/ML
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
timm==0.9.12
segment-anything==1.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
python-dotenv==1.0.1
